analysis-ty: pull blanket db impl into `ty::db`
